rename test vector files to make the them visible to the licence checker.
